[GH-35] GH-35: decompose FileBrowser.tsx (550 lines -> 5 focused modules, all under 150 lines)
